requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
httpx>=0.24.0
python-dotenv>=1.0.0
itsdangerous>=2.1.0
authlib>=1.2.0
aiosqlite>=0.19.0
bcrypt>=4.0.0
jinja2>=3.1.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
